add experimental xml results (12)

+++ api/src/main/java/com/redhat/gss/victims/CheckMate.java
@@ -148,41 +148,60 @@ public class CheckMate {
 
                 ContentDisposition disp = new ContentDisposition(dispString);
                 String fileName = disp.getParameter("filename");
-                if (fileName == null) {
-                    fileName = name;
-                }
 
-                String tmpFileName = null;
-                try {
-                    tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
-                    try {
-                        checkResult.addData(checkOne(db, cache, tmpFileName));
-
-                    } catch (VictimsException e) {
-                        StringBuilder error = new StringBuilder();
-                        error.append("VictimsException while checking:\n");
-                        e.printStackTrace();
-                        error.append(e.toString()).append("\n");
-
-                        CheckResult errorResult = new CheckResult();
-                        errorResult.setTrace(trace.toString());
-                        errorResult.setError(error.toString());
-                        return errorResult;
-                    } catch (IOException e) {
-                        StringBuilder error = new StringBuilder();
-                        error.append("VictimsException while checking:\n");
-                        e.printStackTrace();
-                        error.append(e.toString()).append("\n");
-
-                        CheckResult errorResult = new CheckResult();
-                        errorResult.setTrace(trace.toString());
-                        errorResult.setError(error.toString());
-                        return errorResult;
+                // this is only used for debugging return values from victims
+                if (name.equals("victimsdebug") && fileName == null) {
+                    String s = inputPart.getBodyAsString();
+                    trace.append("victimsdebug: " + s);
+                    String[] a = inputPart.getBodyAsString().split("\\s+");
+                    trace.append(" victimsdebug: " + a);
+                    if (a.length > 0) {
+                        CheckResultElement e = new CheckResultElement();
+                        e.setFile(a[0]);
+                        trace.append(" victimsdebugfile: " + a[0]);
+                        for (int i = 1; i < a.length; i++) {
+                            e.addVulnerability(a[i]);
+                            trace.append(" victimsdebugvuln: " + a[1]);
+                        }
+                        checkResult.addData(e);
+                    }
+                } else {
+                    if (fileName == null) {
+                        fileName = name;
                     }
 
-                } finally {
-                    if (tmpFileName != null) {
-                        deleteTempFile(tmpFileName);
+                    String tmpFileName = null;
+                    try {
+                        tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
+                        try {
+                            checkResult.addData(checkOne(db, cache, tmpFileName));
+
+                        } catch (VictimsException e) {
+                            StringBuilder error = new StringBuilder();
+                            error.append("VictimsException while checking:\n");
+                            e.printStackTrace();
+                            error.append(e.toString()).append("\n");
+
+                            CheckResult errorResult = new CheckResult();
+                            errorResult.setTrace(trace.toString());
+                            errorResult.setError(error.toString());
+                            return errorResult;
+                        } catch (IOException e) {
+                            StringBuilder error = new StringBuilder();
+                            error.append("VictimsException while checking:\n");
+                            e.printStackTrace();
+                            error.append(e.toString()).append("\n");
+
+                            CheckResult errorResult = new CheckResult();
+                            errorResult.setTrace(trace.toString());
+                            errorResult.setError(error.toString());
+                            return errorResult;
+                        }
+
+                    } finally {
+                        if (tmpFileName != null) {
+                            deleteTempFile(tmpFileName);
+                        }
                     }
                 }
             }
@@ -270,16 +289,22 @@ public class CheckMate {
 
                 ContentDisposition disp = new ContentDisposition(dispString);
                 String fileName = disp.getParameter("filename");
-                if (fileName == null) {
-                    fileName = name;
-                }
 
-                String tmpFileName = null;
-                try {
-                    tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
+                // this is only used for debugging return values from victims
+                if (name.equals("victimsdebug") && fileName == null) {
+                    String s = inputPart.getBodyAsString();
+                    trace.append("victimsdebug: " + s);
+                    String[] a = inputPart.getBodyAsString().split("\\s+");
+                    trace.append(" victimsdebug: " + a);
+                    if (a.length > 0) {
+                        CheckResultElement checkResultElement = new CheckResultElement();
+                        checkResultElement.setFile(a[0]);
+                        trace.append(" victimsdebugfile: " + a[0]);
+                        for (int i = 1; i < a.length; i++) {
+                            checkResultElement.addVulnerability(a[i]);
+                            trace.append(" victimsdebugvuln: " + a[1]);
+                        }
 
-                    try {
-                        CheckResultElement checkResultElement = checkOne(db, cache, tmpFileName);
                         String checkFileName = checkResultElement.getFile();
                         List<String> cves = checkResultElement.getVulnerabilities();
                         if (cves != null && cves.size() > 0) {
@@ -292,20 +317,45 @@ public class CheckMate {
                         } else {
                             result.append(checkFileName + " ok\n");
                         }
-
-                    } catch (VictimsException e) {
-                        result.append("VictimsException while checking:\n");
-                        e.printStackTrace();
-                        return result.append(e.toString()).append("\n").toString();
-                    } catch (IOException e) {
-                        result.append("VictimsException while checking:\n");
-                        e.printStackTrace();
-                        return result.append(e.toString()).append("\n").toString();
+                    }
+                } else {
+                    if (fileName == null) {
+                        fileName = name;
                     }
 
-                } finally {
-                    if (tmpFileName != null) {
-                        deleteTempFile(tmpFileName);
+                    String tmpFileName = null;
+                    try {
+                        tmpFileName = copyToTempFile(fileName, inputPart.getBody(InputStream.class, null));
+
+                        try {
+                            CheckResultElement checkResultElement = checkOne(db, cache, tmpFileName);
+                            String checkFileName = checkResultElement.getFile();
+                            List<String> cves = checkResultElement.getVulnerabilities();
+                            if (cves != null && cves.size() > 0) {
+                                result.append(String.format("%s VULNERABLE! ", checkFileName));
+                                for (String cve : cves) {
+                                    result.append(cve);
+                                    result.append(" ");
+                                }
+                                result.append("\n");
+                            } else {
+                                result.append(checkFileName + " ok\n");
+                            }
+
+                        } catch (VictimsException e) {
+                            result.append("VictimsException while checking:\n");
+                            e.printStackTrace();
+                            return result.append(e.toString()).append("\n").toString();
+                        } catch (IOException e) {
+                            result.append("VictimsException while checking:\n");
+                            e.printStackTrace();
+                            return result.append(e.toString()).append("\n").toString();
+                        }
+
+                    } finally {
+                        if (tmpFileName != null) {
+                            deleteTempFile(tmpFileName);
+                        }
                     }
                 }
             }
